requirements.txt
click==8.1.3
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
PyJWT==2.6.0
PyMySQL==1.0.2
python-decouple==3.7
six==1.16.0
SQLAlchemy==2.0.2
typing_extensions==4.4.0
uuid==1.30
Werkzeug==2.2.2
